+++ src/background/requests.ts
@@ -27,7 +27,6 @@ import { ProxiedRemoteServiceError } from "@/services/errors";
 import serviceRegistry from "@/services/registry";
 import { getExtensionToken } from "@/auth/token";
 import { locator } from "@/background/locator";
-import { ContextError, ExtensionNotLinkedError } from "@/errors";
 import { isEmpty } from "lodash";
 import {
   deleteCachedAuthData,
@@ -47,6 +46,7 @@ import {
 import { selectAxiosError } from "@/services/requestErrorUtils";
 import { safeGuessStatusText } from "@/types/errorContract";
 import { BusinessError } from "@/errors/businessErrors";
+import { ContextError, ExtensionNotLinkedError } from "@/errors/genericErrors";
 
 type SanitizedResponse<T = unknown> = Pick<
   AxiosResponse<T>,
