+++ src/runtime/reducePipeline.ts
@@ -26,7 +26,6 @@ import {
   UUID,
 } from "@/core";
 import { castArray, isPlainObject } from "lodash";
-import { ContextError } from "@/errors";
 import {
   clearExtensionDebugLogs,
   getLoggingConfig,
@@ -60,6 +59,7 @@ import { RunBlock } from "@/contentScript/runBlockTypes";
 import { resolveBlockConfig } from "@/blocks/registry";
 import { isObject } from "@/utils";
 import { BusinessError } from "@/errors/businessErrors";
+import { ContextError } from "@/errors/genericErrors";
 
 type CommonOptions = ApiVersionOptions & {
   /**
