+++ src/services/locator.ts
@@ -37,8 +37,8 @@ import {
 import { fetch } from "@/hooks/fetch";
 import { validateRegistryId } from "@/types/helpers";
 import { PIXIEBRIX_SERVICE_ID } from "@/services/constants";
-import { ExtensionNotLinkedError } from "@/errors";
 import { forbidContext } from "@/utils/expectContext";
+import { ExtensionNotLinkedError } from "@/errors/genericErrors";
 
 const REF_SECRETS = [
   "https://app.pixiebrix.com/schemas/key#",
